requirements.txt
ultralytics
opencv-python
Pillow
pyyaml
albumentations>=1.3.0
psutil
GPUtil
